remove dependency on SimpleXMLElement
This commit removes the dependency on `\SimpleXMLElement` for all XPath related
functionality in favor of using `\DOMXPath`.

Also introduced in this commit is a restriction on the `parseDocumentBody()`
method of the plugin to require that all documents have '<!DOCTYPE html' at the
beginning of the document body. This ensures that only documents meant for
client display are processed for HTTP/2 Push, excluding JSON responses, etc.